riscv, soft_isa: support nested emulation for signal handler
Signal handler is an asynchronous event for normal app thread execution,
we cannot directly reassign the in-use vcpu to an asynchronous signal
handler task, that would mess up the vcpu context.

The method introduced here is to assign different vcpu to different task,
so each vcpu's emulation is isolated from each other. Besides, due to
signal handling could be nested, we need kernel to explictly keep a counter
for the nesting level, so soft_isa can conveniently select the
corresponding vcpu for instruction emulation.

TBD:
- For signal handler, the vcpu's initial state would be the ending state of
  a previous signal handler which happened at the same nesting level. This
  normally won't be a issue, since usually software should not read vector
  register before doing vector load.

  If in some rare case this need be addressed, we can inform bt-vdso of info
  like whether it's the first time to serve a certain task. If yes, then the
  previously used vcpu need be reset, otherwise, bt-vdso should continue
  with the vcpu vector state in hand.

- Just as the non-nesting scenario, we don't timely recycle the vcpu when a
  task has finished execution. We expect that a vcpu shouldn't cost much
  memory. Or we can try to recycle the vcpu when a signal handler is
  finished? This would also address the first TBD.

Signed-off-by: Xiao Wang <xiao.w.wang@intel.com>
Signed-off-by: Haicheng Li <haicheng.li@intel.com>