soft_isa: enable dynamic mapping of bt-vdso
To support dynamic mapping of bt-vdso elf image, we need to make the
binary capable of position independent execution (PIE), this requirements
apply to all SW components that contribute to the building process, i.e.
libc, unicorn lib and bt-vdso. Specifically, "-mcmodel=medany" option
should be passed to compiler during the building of all SW components to
eliminate absolute addressing.

For musl lib build
==================
 CROSS_COMPILE=riscv64-linux-gnu- \
 CFLAGS="-mcmodel=medany -mno-relax -fno-pic -O2 -g" \
 ../configure --target=riscv64 --prefix=`pwd`/install/ --disable-shared
 make && make install

Besides, a runtime relocation is needed to adjust the variables of address
type, currently we use a workaround relocation implementation as the glibc
upstream's support of static-pie is WIP.

[haicheng: add $MUSL_BUILD in bt-vdso/Makefile]

Signed-off-by: Xiao Wang <xiao.w.wang@intel.com>
Signed-off-by: Haicheng Li <haicheng.li@intel.com>